New upstream version 2.29.4

+++ Source/WTF/wtf/text/StringParsingBuffer.h
@@ -25,9 +25,7 @@
 
 #pragma once
 
-#include <wtf/Forward.h>
-#include <wtf/text/LChar.h>
-#include <wtf/text/StringCommon.h>
+#include <wtf/text/StringView.h>
 
 namespace WTF {
 
@@ -52,7 +50,7 @@ public:
     {
         ASSERT(characters <= end);
         ASSERT(!characters == !end);
-        ASSERT(end - characters <= std::numeric_limits<unsigned>::max());
+        ASSERT(static_cast<size_t>(end - characters) <= std::numeric_limits<unsigned>::max());
     }
 
     constexpr auto position() const { return m_position; }
